feat(agent): relationship-first get_timeline format
- Clips report frames: [start, end) instead of startFrame/durationFrames/
  endFrame — adjacency reads directly and matches gaps and ripple ranges.
- Linked visual+audio pairs fold into the visual clip as audio: {id,
  track, ...deviations}; the partner track reports linkedClips instead of
  repeating mirror clips, and linkGroupId leaves the payload.
- Keyframe tracks that animate nothing collapse: identity tracks are
  dropped, constant tracks become the static field they emulate (values
  compared at display precision).
- Caption tracks no longer report pause gaps; empty clips arrays are
  omitted; all tool JSON is emitted with sorted keys.
- get_media/inspect_media move to ToolExecutor+Media.swift; the timeline
  file keeps its tools plus the payload-shaping helpers.
