+++ Sources/PalmierPro/Agent/Tools/ToolExecutor+Timeline.swift
@@ -1,47 +1,20 @@
-import AVFoundation
 import Foundation
-import ImageIO
 
+// get_timeline, create_timeline, set_active_timeline.
 extension ToolExecutor {
-    private static let defaultReadVideoFrames = 6
-    private static let readVideoMaxFrames = 12
-    private nonisolated static let readVideoFrameMaxDimension: CGFloat = 512
-    private nonisolated static let readVideoJPEGQuality: CGFloat = 0.7
-
     private static let getTimelineAllowedKeys: Set<String> = ["startFrame", "endFrame", "captionDetail"]
-    private static let captionRowLimit = 200
-    private static let captionRowFormat = ["clipId", "startFrame", "durationFrames", "text"]
-    private static let captionPreviewLimit = 60
 
     func getTimeline(_ editor: EditorViewModel, _ args: [String: Any]) throws -> ToolResult {
         try validateUnknownKeys(args, allowed: Self.getTimelineAllowedKeys, path: "get_timeline")
-        var window: Range<Int>?
-        if args.int("startFrame") != nil || args.int("endFrame") != nil {
-            let s = args.int("startFrame") ?? 0
-            let e = args.int("endFrame") ?? Int.max
-            guard s < e else {
-                throw ToolError("Invalid window [\(s), \(e)): startFrame must be less than endFrame")
-            }
-            window = s..<e
-        }
+        let window = try Self.frameWindow(args)
         let captionDetail = args["captionDetail"] as? Bool ?? false
 
         guard var dict = try? JSONSerialization.jsonObject(
             with: JSONEncoder().encode(editor.timeline)
         ) as? [String: Any] else { throw ToolError("Failed to encode timeline") }
         dict.removeValue(forKey: "settingsConfigured")
-        if var tracks = dict["tracks"] as? [[String: Any]] {
-            for i in tracks.indices {
-                tracks[i] = Self.compactTrack(tracks[i], window: window, captionDetail: captionDetail)
-                // Report the displayed label (mirrored video numbering), not the stored seed.
-                tracks[i]["label"] = editor.timelineTrackDisplayLabel(at: i)
-                tracks[i]["index"] = i
-                tracks[i].removeValue(forKey: "id")
-                tracks[i].removeValue(forKey: "displayHeight")
-                let gaps = Self.trackGaps(editor.timeline.tracks[i])
-                if !gaps.isEmpty { tracks[i]["gaps"] = gaps }
-            }
-            dict["tracks"] = tracks
+        if let tracks = dict["tracks"] as? [[String: Any]] {
+            dict["tracks"] = Self.compactTracks(tracks, editor: editor, window: window, captionDetail: captionDetail)
         }
         dict["totalFrames"] = editor.timeline.totalFrames
         dict["durationSeconds"] = Double(editor.timeline.totalFrames) / Double(max(editor.timeline.fps, 1))
@@ -61,6 +34,16 @@ extension ToolExecutor {
             throw ToolError("Failed to encode timeline")
         }
         return .ok(json)
+    }
+
+    private static func frameWindow(_ args: [String: Any]) throws -> Range<Int>? {
+        guard args.int("startFrame") != nil || args.int("endFrame") != nil else { return nil }
+        let s = args.int("startFrame") ?? 0
+        let e = args.int("endFrame") ?? Int.max
+        guard s < e else {
+            throw ToolError("Invalid window [\(s), \(e)): startFrame must be less than endFrame")
+        }
+        return s..<e
     }
 
     func createTimeline(_ editor: EditorViewModel, _ args: [String: Any]) throws -> ToolResult {
@@ -113,6 +96,12 @@ extension ToolExecutor {
         return .ok(Self.jsonString(payload) ?? "{}")
     }
 
+    // MARK: - Payload shaping
+
+    private static let captionRowLimit = 200
+    private static let captionRowFormat = ["clipId", "startFrame", "endFrame", "text"]
+    private static let captionPreviewLimit = 60
+
     private static let trackDefaults: [String: Any] = ["muted": false, "hidden": false, "syncLocked": true]
 
     private static let clipDefaults: [String: Any] = {
@@ -127,9 +116,28 @@ extension ToolExecutor {
         return obj
     }()
 
-    /// Internal empty spans between clips; leading/trailing space is readable off the clip frames.
+    private static func compactTracks(
+        _ tracks: [[String: Any]], editor: EditorViewModel, window: Range<Int>?, captionDetail: Bool
+    ) -> [[String: Any]] {
+        let fold = linkFoldPlan(tracks)
+        return tracks.indices.map { i in
+            var track = compactTrack(tracks[i], window: window, captionDetail: captionDetail, fold: fold)
+            // Report the displayed label (mirrored video numbering), not the stored seed.
+            track["label"] = editor.timelineTrackDisplayLabel(at: i)
+            track["index"] = i
+            track.removeValue(forKey: "id")
+            track.removeValue(forKey: "displayHeight")
+            if let count = fold.foldedCountByTrack[i] { track["linkedClips"] = count }
+            let gaps = trackGaps(editor.timeline.tracks[i])
+            if !gaps.isEmpty { track["gaps"] = gaps }
+            return track
+        }
+    }
+
+    /// Empty spans between non-caption clips; leading/trailing space reads off the clip frames.
     private static func trackGaps(_ track: Track) -> [[Int]] {
         let spans = track.clips
+            .filter { $0.captionGroupId == nil }
             .map { [$0.startFrame, $0.startFrame + $0.durationFrames] }
             .sorted { $0[0] < $1[0] }
         var gaps: [[Int]] = []
@@ -141,10 +149,69 @@ extension ToolExecutor {
         return gaps
     }
 
-    private static func compactTrack(_ track: [String: Any], window: Range<Int>?, captionDetail: Bool) -> [String: Any] {
+    // MARK: - Linked audio fold
+
+    /// Visual+audio linked pairs fold into the visual clip; linkage reads structurally, not via linkGroupId.
+    struct LinkFold {
+        var partnerByVisualId: [String: (clip: [String: Any], trackIndex: Int)] = [:]
+        var foldedAudioIds: Set<String> = []
+        var foldedCountByTrack: [Int: Int] = [:]
+    }
+
+    private static func linkFoldPlan(_ tracks: [[String: Any]]) -> LinkFold {
+        var groups: [String: [(clip: [String: Any], trackIndex: Int)]] = [:]
+        for (ti, track) in tracks.enumerated() {
+            for clip in track["clips"] as? [[String: Any]] ?? [] {
+                if let gid = clip["linkGroupId"] as? String {
+                    groups[gid, default: []].append((clip, ti))
+                }
+            }
+        }
+        var plan = LinkFold()
+        for members in groups.values where members.count == 2 {
+            guard let audio = members.first(where: { ($0.clip["mediaType"] as? String) == "audio" }),
+                  let visual = members.first(where: { ($0.clip["mediaType"] as? String) != "audio" }),
+                  let visualId = visual.clip["id"] as? String,
+                  let audioId = audio.clip["id"] as? String else { continue }
+            plan.partnerByVisualId[visualId] = (audio.clip, audio.trackIndex)
+            plan.foldedAudioIds.insert(audioId)
+            plan.foldedCountByTrack[audio.trackIndex, default: 0] += 1
+        }
+        return plan
+    }
+
+    /// The folded audio partner: id, track, and only what deviates from its visual clip.
+    private static func audioSummary(
+        _ partner: [String: Any], trackIndex: Int, visual: [String: Any]
+    ) -> [String: Any] {
+        var out: [String: Any] = ["id": partner["id"] ?? "", "track": trackIndex]
+        let pStart = intValue(partner["startFrame"])
+        let pDur = intValue(partner["durationFrames"])
+        if pStart != intValue(visual["startFrame"]) || pDur != intValue(visual["durationFrames"]) {
+            out["frames"] = [pStart, pStart + pDur]
+        }
+        for key in ["trimStartFrame", "trimEndFrame", "speed"] {
+            if let pv = partner[key] as? NSNumber, let vv = visual[key] as? NSNumber, pv != vv {
+                out[key] = pv
+            }
+        }
+        let stripped = strippingDefaults(compactClipKeyframes(partner), clipDefaults)
+        for key in ["volume", "fadeInFrames", "fadeOutFrames", "fadeInInterpolation", "fadeOutInterpolation", "effects", "keyframes"] {
+            if let v = stripped[key] { out[key] = v }
+        }
+        return out
+    }
+
+    // MARK: - Track and clip compaction
+
+    private static func compactTrack(
+        _ track: [String: Any], window: Range<Int>?, captionDetail: Bool, fold: LinkFold
+    ) -> [String: Any] {
         var out = strippingDefaults(track, trackDefaults)
         guard let rawClips = track["clips"] as? [[String: Any]] else { return out }
-        let compacted = rawClips.map { compactClip($0) }
+        let compacted = rawClips
+            .filter { !fold.foldedAudioIds.contains(($0["id"] as? String) ?? "") }
+            .map { compactClip($0, fold: fold) }
 
         var loose: [[String: Any]] = []
         var groupOrder: [String] = []
@@ -164,21 +231,17 @@ extension ToolExecutor {
             groups.append(group)
             loose.append(contentsOf: deviants)
         }
-        loose = loose.map { clip in
-            var c = clip
-            c["endFrame"] = intValue(c["startFrame"]) + intValue(c["durationFrames"])
-            return c
-        }
-        loose.sort { intValue($0["startFrame"]) < intValue($1["startFrame"]) }
+        loose.sort { intValue(($0["frames"] as? [Any])?.first) < intValue(($1["frames"] as? [Any])?.first) }
 
         let visible = window.map { w in loose.filter { clipIntersects($0, w) } } ?? loose
-        out["clips"] = visible
+        out.removeValue(forKey: "clips")
+        if !visible.isEmpty { out["clips"] = visible }
         if visible.count < loose.count { out["totalClips"] = loose.count }
         if !groups.isEmpty { out["captionGroups"] = groups }
         return out
     }
 
-    private static func compactClip(_ clip: [String: Any]) -> [String: Any] {
+    private static func compactClip(_ clip: [String: Any], fold: LinkFold? = nil) -> [String: Any] {
         var out = compactClipKeyframes(clip)
         if let s = out["sourceClipType"] as? String, s == out["mediaType"] as? String {
             out.removeValue(forKey: "sourceClipType")
@@ -188,7 +251,16 @@ extension ToolExecutor {
             out.removeValue(forKey: "trimStartFrame")
             out.removeValue(forKey: "trimEndFrame")
         }
-        return strippingDefaults(out, clipDefaults)
+        out = strippingDefaults(out, clipDefaults)
+        let start = intValue(out["startFrame"])
+        out["frames"] = [start, start + intValue(out["durationFrames"])]
+        out.removeValue(forKey: "startFrame")
+        out.removeValue(forKey: "durationFrames")
+        if let fold, let id = out["id"] as? String, let partner = fold.partnerByVisualId[id] {
+            out["audio"] = audioSummary(partner.clip, trackIndex: partner.trackIndex, visual: clip)
+            out.removeValue(forKey: "linkGroupId")
+        }
+        return out
     }
 
     /// Removes keys whose values equal the defaults; recurses into nested objects.
@@ -206,11 +278,13 @@ extension ToolExecutor {
         return out
     }
 
+    // MARK: - Caption groups
+
     /// Collapses one caption group into shared properties + a summary (default) or compact rows (detail).
     private static func captionGroup(
         gid: String, members: [[String: Any]], window: Range<Int>?, detail: Bool
     ) -> (group: [String: Any], deviants: [[String: Any]]) {
-        let rowKeys: Set<String> = ["id", "startFrame", "durationFrames", "textContent", "captionGroupId"]
+        let rowKeys: Set<String> = ["id", "frames", "textContent", "captionGroupId"]
         var counts: [String: Int] = [:]
         var modalKey = ""
         var shared: [String: Any] = [:]
@@ -236,12 +310,13 @@ extension ToolExecutor {
         var frameMin = Int.max
         var frameMax = 0
         for (clip, key) in entries {
-            let start = intValue(clip["startFrame"])
-            let end = start + intValue(clip["durationFrames"])
+            let frames = clip["frames"] as? [Any]
+            let start = intValue(frames?.first)
+            let end = intValue(frames?.last)
             frameMin = min(frameMin, start)
             frameMax = max(frameMax, end)
             if key == modalKey {
-                rows.append([clip["id"] ?? "", start, end - start, clip["textContent"] ?? ""])
+                rows.append([clip["id"] ?? "", start, end, clip["textContent"] ?? ""])
             } else {
                 deviants.append(clip)
             }
@@ -249,7 +324,7 @@ extension ToolExecutor {
 
         let total = rows.count
         if let window {
-            rows = rows.filter { intValue($0[1]) < window.upperBound && intValue($0[1]) + intValue($0[2]) > window.lowerBound }
+            rows = rows.filter { intValue($0[1]) < window.upperBound && intValue($0[2]) > window.lowerBound }
         }
         rows.sort { intValue($0[1]) < intValue($1[1]) }
 
@@ -283,17 +358,7 @@ extension ToolExecutor {
         text.count > captionPreviewLimit ? String(text.prefix(captionPreviewLimit)) + "…" : text
     }
 
-    private static func canonicalJSON(_ obj: [String: Any]) -> String {
-        guard let data = try? JSONSerialization.data(withJSONObject: obj, options: [.sortedKeys]) else { return "" }
-        return String(data: data, encoding: .utf8) ?? ""
-    }
-
-    private static func clipIntersects(_ clip: [String: Any], _ window: Range<Int>) -> Bool {
-        let start = intValue(clip["startFrame"])
-        return start < window.upperBound && start + intValue(clip["durationFrames"]) > window.lowerBound
-    }
-
-    private static func intValue(_ v: Any?) -> Int { (v as? NSNumber)?.intValue ?? 0 }
+    // MARK: - Keyframes
 
     private static func compactClipKeyframes(_ clip: [String: Any]) -> [String: Any] {
         var out = clip
@@ -306,22 +371,59 @@ extension ToolExecutor {
             ("scaleTrack", "scale", KeyframeValueShape.pair),
             ("cropTrack", "crop", KeyframeValueShape.crop),
         ] {
-            if let track = clip[trackKey] as? [String: Any],
-               let kfs = track["keyframes"] as? [[String: Any]],
-               !kfs.isEmpty {
-                keyframes[propKey] = kfs.map { kf -> [Any] in
-                    var row: [Any] = [kf["frame"] ?? 0]
-                    row.append(contentsOf: valueShape.values(from: kf["value"]))
-                    if let interp = kf["interpolationOut"] as? String, interp != "smooth" {
-                        row.append(interp)
-                    }
-                    return row
-                }
+            defer { out.removeValue(forKey: trackKey) }
+            guard let track = clip[trackKey] as? [String: Any],
+                  let kfs = track["keyframes"] as? [[String: Any]],
+                  !kfs.isEmpty else { continue }
+
+            let values = kfs.map { valueShape.values(from: $0["value"]).map { ($0 as? NSNumber)?.doubleValue ?? 0 } }
+            if let first = values.first, values.allSatisfy({ nearlyEqual($0, first) }),
+               collapseConstantKeyframes(first, propKey: propKey, clip: clip, into: &out) {
+                continue
             }
-            out.removeValue(forKey: trackKey)
+
+            keyframes[propKey] = kfs.map { kf -> [Any] in
+                var row: [Any] = [kf["frame"] ?? 0]
+                row.append(contentsOf: valueShape.values(from: kf["value"]))
+                if let interp = kf["interpolationOut"] as? String, interp != "smooth" {
+                    row.append(interp)
+                }
+                return row
+            }
         }
         if !keyframes.isEmpty { out["keyframes"] = keyframes }
         return out
+    }
+
+    /// True when absorbed: identity tracks vanish; constants become the static field when it's at default.
+    private static func collapseConstantKeyframes(
+        _ value: [Double], propKey: String, clip: [String: Any], into out: inout [String: Any]
+    ) -> Bool {
+        switch propKey {
+        case "volume", "opacity":
+            if nearlyEqual(value, [1]) { return true }
+            guard (clip[propKey] as? NSNumber)?.doubleValue == 1 else { return false }
+            out[propKey] = value[0]
+            return true
+        case "rotation":
+            return nearlyEqual(value, [0])
+        case "position":
+            return nearlyEqual(value, [0, 0])
+        case "scale":
+            return nearlyEqual(value, [1, 1])
+        case "crop":
+            if nearlyEqual(value, [0, 0, 0, 0]) { return true }
+            let staticCrop = clip["crop"] as? [String: Any] ?? [:]
+            guard staticCrop.values.allSatisfy({ (($0 as? NSNumber)?.doubleValue ?? 0) == 0 }) else { return false }
+            out["crop"] = ["top": value[0], "right": value[1], "bottom": value[2], "left": value[3]]
+            return true
+        default:
+            return false
+        }
+    }
+
+    private static func nearlyEqual(_ a: [Double], _ b: [Double]) -> Bool {
+        a.count == b.count && zip(a, b).allSatisfy { abs($0 - $1) < 0.0005 }
     }
 
     private enum KeyframeValueShape {
@@ -341,344 +443,17 @@ extension ToolExecutor {
         }
     }
 
-    private static let getMediaAllowedKeys: Set<String> = ["ids", "folder", "pending"]
-    private static let getMediaPromptLimit = 100
+    // MARK: - Small helpers
 
-    func getMedia(_ editor: EditorViewModel, _ args: [String: Any]) throws -> ToolResult {
-        try validateUnknownKeys(args, allowed: Self.getMediaAllowedKeys, path: "get_media")
-        let idFilter = Set(args.stringArray("ids"))
-        let pendingOnly = args["pending"] as? Bool ?? false
-        var folderScope: Set<String>?
-        if let folderPath = args.string("folder") {
-            folderScope = folderIdsIncludingDescendants(
-                [try folderId(atPath: folderPath, editor: editor)], editor: editor)
-        }
-
-        var assets: [[String: Any]] = []
-        for entry in editor.mediaManifest.entries {
-            if !idFilter.isEmpty && !idFilter.contains(entry.id) { continue }
-            let status = entry.generationStatus
-            let pending = status != nil && status != "none"
-            if pendingOnly && !pending { continue }
-            if let folderScope {
-                guard let f = entry.folderId, folderScope.contains(f) else { continue }
-            }
-            var a: [String: Any] = ["id": entry.id, "name": entry.name, "type": entry.type.rawValue]
-            if entry.duration > 0 { a["durationSeconds"] = entry.duration }
-            if let w = entry.sourceWidth, let h = entry.sourceHeight { a["width"] = w; a["height"] = h }
-            if let fps = entry.sourceFPS { a["fps"] = fps }
-            if entry.hasAudio == true, entry.type == .video { a["hasAudio"] = true }
-            if let path = folderPathString(entry.folderId, editor: editor) { a["folder"] = path }
-            if pending, let status { a["generationStatus"] = status }
-            if let prompt = entry.generationInput?.prompt, !prompt.isEmpty {
-                a["prompt"] = prompt.count > Self.getMediaPromptLimit
-                    ? String(prompt.prefix(Self.getMediaPromptLimit)) + "…" : prompt
-            }
-            assets.append(a)
-        }
-
-        var payload: [String: Any] = ["assets": assets]
-        // Full inventory only on unfiltered reads; filtered reads (polling) stay minimal.
-        if idFilter.isEmpty && !pendingOnly && folderScope == nil {
-            let folderPaths = allFolderPaths(editor)
-            if !folderPaths.isEmpty { payload["folders"] = folderPaths }
-            payload["timelines"] = editor.timelines.map { t -> [String: Any] in
-                var e: [String: Any] = [
-                    "timelineId": t.id, "name": t.name,
-                    "durationSeconds": Double(t.totalFrames) / Double(max(t.fps, 1)),
-                ]
-                if t.id == editor.activeTimelineId { e["active"] = true }
-                if let path = folderPathString(t.folderId, editor: editor) { e["folder"] = path }
-                return e
-            }
-        }
-        guard let json = Self.jsonString(roundJSONFloatingPointNumbers(payload, toPlaces: 3)) else {
-            throw ToolError("Failed to encode media library")
-        }
-        return .ok(json)
+    private static func canonicalJSON(_ obj: [String: Any]) -> String {
+        guard let data = try? JSONSerialization.data(withJSONObject: obj, options: [.sortedKeys]) else { return "" }
+        return String(data: data, encoding: .utf8) ?? ""
     }
 
-    private static let inspectMediaAllowedKeys: Set<String> = [
-        "mediaRef", "clipId", "maxFrames", "startSeconds", "endSeconds", "wordTimestamps", "overview", "language",
-    ]
-
-    func inspectMedia(_ editor: EditorViewModel, _ args: [String: Any]) async throws -> ToolResult {
-        try validateUnknownKeys(args, allowed: Self.inspectMediaAllowedKeys, path: "inspect_media")
-        let mediaRef = try args.requireString("mediaRef")
-        let asset = try asset(mediaRef, editor: editor)
-        let url = asset.url
-        guard FileManager.default.fileExists(atPath: url.path) else {
-            switch asset.generationStatus {
-            case .preparing:
-                throw ToolError("Asset \(asset.id) is still preparing. Poll get_media and retry once generationStatus becomes 'none'.")
-            case .downloading:
-                throw ToolError("Asset \(asset.id) is still downloading. Poll get_media and retry once generationStatus becomes 'none'.")
-            case .generating:
-                throw ToolError("Asset \(asset.id) is still generating. Poll get_media and retry once generationStatus becomes 'none'.")
-            case .rendering:
-                throw ToolError("Asset \(asset.id) is still rendering. Poll get_media and retry once generationStatus becomes 'none'.")
-            case .failed(let msg):
-                throw ToolError("Asset \(asset.id) failed: \(msg)")
-            case .none:
-                throw ToolError("Media file not on disk: \(url.lastPathComponent)")
-            }
-        }
-
-        var mapping: (clip: Clip, fps: Int)?
-        if let clipId = args.string("clipId") {
-            guard let loc = editor.findClip(id: clipId) else {
-                throw ToolError("Clip not found: \(clipId)")
-            }
-            let clip = editor.timeline.tracks[loc.trackIndex].clips[loc.clipIndex]
-            guard clip.mediaRef == mediaRef else {
-                throw ToolError("Clip \(clipId) does not reference mediaRef \(mediaRef) (it references \(clip.mediaRef))")
-            }
-            mapping = (clip, editor.timeline.fps)
-        }
-
-        let preferredLocale = try await Self.parseLocale(args, path: "inspect_media")
-
-        switch asset.type {
-        case .image: return try await readImage(asset: asset, args: args)
-        case .video: return try await readVideo(editor: editor, asset: asset, args: args, mapping: mapping, preferredLocale: preferredLocale)
-        case .audio: return try await readAudio(editor: editor, asset: asset, args: args, mapping: mapping, preferredLocale: preferredLocale)
-        case .lottie: return try await readLottie(asset: asset, args: args)
-        case .text: throw ToolError("Text clips are not stored as media assets.")
-        case .sequence: throw ToolError("Sequences are timelines, not media assets. Use get_timeline.")
-        }
+    private static func clipIntersects(_ clip: [String: Any], _ window: Range<Int>) -> Bool {
+        let frames = clip["frames"] as? [Any]
+        return intValue(frames?.first) < window.upperBound && intValue(frames?.last) > window.lowerBound
     }
 
-    private static func sourceRange(_ args: [String: Any], duration: Double) throws -> ClosedRange<Double>? {
-        let start = args.double("startSeconds")
-        let end = args.double("endSeconds")
-        guard start != nil || end != nil else { return nil }
-        let s = max(start ?? 0, 0)
-        let e = min(end ?? duration, duration)
-        guard s < e else {
-            throw ToolError("Invalid time range [\(s), \(e)] for media of duration \(duration)s")
-        }
-        return s...e
-    }
-
-    private func readImage(asset: MediaAsset, args: [String: Any]) async throws -> ToolResult {
-        let url = asset.url
-        let encoded = await Task.detached(priority: .userInitiated) {
-            ImageEncoder.encode(url: url).map {
-                (base64: $0.data.base64EncodedString(), mime: $0.mime, encodedByteSize: $0.data.count)
-            }
-        }.value
-        guard let encoded else {
-            throw ToolError("Failed to read or decode image file")
-        }
-
-        let fileSize = (try? FileManager.default.attributesOfItem(atPath: url.path)[.size] as? NSNumber)?.uint64Value ?? 0
-        var meta = Self.baseMeta(for: asset)
-        meta["mimeType"] = encoded.mime
-        meta["byteSize"] = fileSize
-        meta["encodedByteSize"] = encoded.encodedByteSize
-        if let props = Self.imagePropertiesSummary(at: url) {
-            meta["imageProperties"] = props
-        }
-
-        guard let metaJSON = Self.jsonString(roundJSONFloatingPointNumbers(meta, toPlaces: 3)) else {
-            throw ToolError("Failed to encode metadata")
-        }
-        return ToolResult(
-            content: [.image(base64: encoded.base64, mediaType: encoded.mime), .text(metaJSON)],
-            isError: false
-        )
-    }
-
-    private func readVideo(editor: EditorViewModel, asset: MediaAsset, args: [String: Any], mapping: (clip: Clip, fps: Int)? = nil, preferredLocale: Locale? = nil) async throws -> ToolResult {
-        guard asset.duration > 0 else { throw ToolError("Video has zero duration: \(asset.name)") }
-
-        let range = try Self.sourceRange(args, duration: asset.duration)
-        let windowStart = range?.lowerBound ?? 0
-        let windowEnd = range?.upperBound ?? asset.duration
-
-        var meta = Self.baseMeta(for: asset)
-        meta["hasAudio"] = asset.hasAudio
-        if let range { meta["timeRange"] = [range.lowerBound, range.upperBound] }
-
-        // Frames/overview and transcription touch independent subsystems — run them concurrently
-        let url = asset.url
-        let hasAudio = asset.hasAudio
-        let wantsOverview = args.bool("overview") == true
-        let requested = args.int("maxFrames") ?? Self.defaultReadVideoFrames
-        let frameCount = max(1, min(requested, Self.readVideoMaxFrames))
-        async let visualTask = Self.extractVisual(
-            url: url, name: asset.name, overview: wantsOverview,
-            frameCount: frameCount, start: windowStart, end: windowEnd
-        )
-        async let transcriptTask: Result<TranscriptionResult, Error>? = {
-            guard hasAudio else { return nil }
-            do { return .success(try await TranscriptCache.shared.transcript(for: url, isVideo: true, range: range, preferredLocale: preferredLocale)) }
-            catch { return .failure(error) }
-        }()
-
-        var imageBlocks: [ToolResult.Block] = []
-        switch try await visualTask {
-        case .overview(let jpeg, let timestamps):
-            meta["overview"] = ["tileTimestamps": timestamps.map { $0.jsonRounded(toPlaces: 3) }]
-            imageBlocks = [.image(base64: jpeg.base64EncodedString(), mediaType: "image/jpeg")]
-        case .frames(let frames):
-            meta["frameTimestamps"] = frames.map { $0.timestamp.jsonRounded(toPlaces: 3) }
-            imageBlocks = frames.map { .image(base64: $0.jpeg.base64EncodedString(), mediaType: "image/jpeg") }
-        }
-
-        switch await transcriptTask {
-        case .success(let transcript):
-            meta["transcription"] = Self.transcriptionMeta(
-                from: transcript, mapping: mapping, includeWords: args.bool("wordTimestamps") ?? false
-            )
-        case .failure(let error):
-            Log.transcription.error("video transcription failed: \(error.localizedDescription)")
-            meta["transcriptionError"] = error.localizedDescription
-        case nil:
-            break
-        }
-        if let mapping { meta["timelineMapping"] = Self.timelineMappingMeta(clip: mapping.clip, fps: mapping.fps) }
-
-        guard let metaJSON = Self.jsonString(roundJSONFloatingPointNumbers(meta, toPlaces: 3)) else {
-            throw ToolError("Failed to encode metadata")
-        }
-        return ToolResult(content: imageBlocks + [.text(metaJSON)], isError: false)
-    }
-
-    private enum Visual: Sendable {
-        case frames([(timestamp: Double, jpeg: Data)])
-        case overview(jpeg: Data, timestamps: [Double])
-    }
-
-    private nonisolated static func extractVisual(
-        url: URL, name: String, overview: Bool, frameCount: Int, start: Double, end: Double
-    ) async throws -> Visual {
-        if overview {
-            do {
-                let sheet = try await OverviewRenderer.make(url: url, start: start, end: end)
-                return .overview(jpeg: sheet.jpeg, timestamps: sheet.timestamps)
-            } catch {
-                throw ToolError("Overview failed: \(error.localizedDescription)")
-            }
-        }
-
-        let asset = AVURLAsset(url: url)
-        guard (try? await asset.loadTracks(withMediaType: .video).first) != nil else {
-            throw ToolError("No video track available in \(name)")
-        }
-        let generator = AVAssetImageGenerator(asset: asset)
-        generator.appliesPreferredTrackTransform = true
-        generator.maximumSize = CGSize(
-            width: readVideoFrameMaxDimension,
-            height: readVideoFrameMaxDimension
-        )
-        generator.requestedTimeToleranceBefore = CMTime(seconds: 0.25, preferredTimescale: 600)
-        generator.requestedTimeToleranceAfter = CMTime(seconds: 0.25, preferredTimescale: 600)
-
-        var frames: [(timestamp: Double, jpeg: Data)] = []
-        for i in 0..<frameCount {
-            let t = start + (end - start) * (Double(i) + 0.5) / Double(frameCount)
-            let cmTime = CMTime(seconds: t, preferredTimescale: 600)
-            guard let cgImage = try? await generator.image(at: cmTime).image else { continue }
-            guard let jpeg = ImageEncoder.encodeJPEG(cgImage, quality: readVideoJPEGQuality) else { continue }
-            frames.append((timestamp: t, jpeg: jpeg))
-        }
-        guard !frames.isEmpty else { throw ToolError("Failed to extract frames from \(name)") }
-        return .frames(frames)
-    }
-
-    private func readLottie(asset: MediaAsset, args: [String: Any]) async throws -> ToolResult {
-        let count = max(1, min(args.int("maxFrames") ?? Self.defaultReadVideoFrames, Self.readVideoMaxFrames))
-        let (lottieMeta, frames) = try await LottieVideoGenerator.sampleFrames(fileAt: asset.url, count: count)
-        guard !frames.isEmpty else { throw ToolError("Failed to render Lottie frames from \(asset.name)") }
-
-        var meta = Self.baseMeta(for: asset)
-        meta["framerate"] = lottieMeta.framerate
-        meta["frameCount"] = lottieMeta.frameCount
-        meta["durationSeconds"] = lottieMeta.duration
-        meta["sampledFrameIndices"] = frames.map(\.frameIndex)
-        meta["note"] = "Lottie frames sampled evenly across the animation; transparent areas composited over gray."
-
-        let imageBlocks: [ToolResult.Block] = frames.compactMap { frame in
-            Self.compositeJPEG(frame.image).map { .image(base64: $0.base64EncodedString(), mediaType: "image/jpeg") }
-        }
-        guard !imageBlocks.isEmpty else { throw ToolError("Failed to encode Lottie frames") }
-        guard let metaJSON = Self.jsonString(roundJSONFloatingPointNumbers(meta, toPlaces: 3)) else {
-            throw ToolError("Failed to encode metadata")
-        }
-        return ToolResult(content: imageBlocks + [.text(metaJSON)], isError: false)
-    }
-
-    /// Composites an alpha frame over mid-gray so transparent regions read clearly to the model.
-    private static func compositeJPEG(_ image: CGImage, quality: CGFloat = 0.7) -> Data? {
-        guard let context = CGContext(
-            data: nil, width: image.width, height: image.height,
-            bitsPerComponent: 8, bytesPerRow: 0,
-            space: CGColorSpace(name: CGColorSpace.sRGB) ?? CGColorSpaceCreateDeviceRGB(),
-            bitmapInfo: CGImageAlphaInfo.noneSkipLast.rawValue
-        ) else { return nil }
-        let rect = CGRect(x: 0, y: 0, width: image.width, height: image.height)
-        context.setFillColor(gray: 0.5, alpha: 1)
-        context.fill(rect)
-        context.draw(image, in: rect)
-        return context.makeImage().flatMap { ImageEncoder.encodeJPEG($0, quality: quality) }
-    }
-
-    private func readAudio(editor: EditorViewModel, asset: MediaAsset, args: [String: Any], mapping: (clip: Clip, fps: Int)? = nil, preferredLocale: Locale? = nil) async throws -> ToolResult {
-        let range = try Self.sourceRange(args, duration: asset.duration)
-        let transcript: TranscriptionResult
-        do {
-            transcript = try await TranscriptCache.shared.transcript(for: asset.url, isVideo: false, range: range, preferredLocale: preferredLocale)
-        } catch {
-            throw ToolError("Transcription failed: \(error.localizedDescription)")
-        }
-
-        var meta = Self.baseMeta(for: asset)
-        if let range { meta["timeRange"] = [range.lowerBound, range.upperBound] }
-        let transcription = Self.transcriptionMeta(
-            from: transcript, mapping: mapping, includeWords: args.bool("wordTimestamps") ?? false
-        )
-        for (k, v) in transcription { meta[k] = v }
-        if let mapping { meta["timelineMapping"] = Self.timelineMappingMeta(clip: mapping.clip, fps: mapping.fps) }
-        guard let metaJSON = Self.jsonString(roundJSONFloatingPointNumbers(meta, toPlaces: 3)) else {
-            throw ToolError("Failed to encode metadata")
-        }
-        return .ok(metaJSON)
-    }
-
-    private static func baseMeta(for asset: MediaAsset) -> [String: Any] {
-        var meta: [String: Any] = [
-            "id": asset.id, "name": asset.name,
-            "type": asset.type.rawValue, "duration": asset.duration.jsonRounded(toPlaces: 3),
-            "fileName": asset.url.lastPathComponent,
-            "generationStatus": asset.generationStatus.serialized,
-        ]
-        if let w = asset.sourceWidth { meta["sourceWidth"] = w }
-        if let h = asset.sourceHeight { meta["sourceHeight"] = h }
-        if let fps = asset.sourceFPS { meta["sourceFPS"] = fps }
-        if let gi = asset.generationInput, let obj = encodeAsJSONObject(gi) {
-            meta["generationInput"] = obj
-        }
-        return meta
-    }
-
-    private static func encodeAsJSONObject<T: Encodable>(_ value: T) -> Any? {
-        guard let data = try? JSONEncoder().encode(value),
-              let obj = try? JSONSerialization.jsonObject(with: data)
-        else { return nil }
-        return obj
-    }
-
-    private static func imagePropertiesSummary(at url: URL) -> [String: Any]? {
-        guard let source = CGImageSourceCreateWithURL(url as CFURL, nil),
-              let props = CGImageSourceCopyPropertiesAtIndex(source, 0, nil) as? [CFString: Any]
-        else { return nil }
-        var out: [String: Any] = [:]
-        if let v = props[kCGImagePropertyPixelWidth] { out["pixelWidth"] = v }
-        if let v = props[kCGImagePropertyPixelHeight] { out["pixelHeight"] = v }
-        if let v = props[kCGImagePropertyOrientation] { out["orientation"] = v }
-        if let v = props[kCGImagePropertyDepth] { out["depth"] = v }
-        if let v = props[kCGImagePropertyColorModel] { out["colorModel"] = v }
-        return out.isEmpty ? nil : out
-    }
+    private static func intValue(_ v: Any?) -> Int { (v as? NSNumber)?.intValue ?? 0 }
 }
